API: evict station name and PRN from dense data
These are constant values in these arrays and there is always a structure above
the numpy array that is carrying this information, we can evict the strings
from the arrays.

A key benefit of doing this (other than the memory savings) is that now it is
trivial to write the structured data arrays to hdf5 files.